# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMONTA BENCANA API Documentation",
        "description": "Sistem Informasi Manajemen Bencana Terpadu - RESTful API untuk pengelolaan data bencana, laporan, monitoring, dan integrasi dengan BMKG.",
        "contact": {
            "email": "support@simonta-bencana.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User Baru",
                "description": "Endpoint untuk mendaftarkan user baru ke dalam sistem SIMONTA BENCANA.",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "username",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "nama": {
                                        "description": "Nama lengkap user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "description": "Username unik untuk login",
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "description": "Password minimal 6 karakter",
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "role": {
                                        "description": "Role user",
                                        "type": "string",
                                        "enum": [
                                            "Admin",
                                            "PetugasBPBD",
                                            "OperatorDesa",
                                            "Warga"
                                        ],
                                        "example": "Warga"
                                    },
                                    "email": {
                                        "description": "Email user",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "no_telepon": {
                                        "description": "Nomor telepon",
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "alamat": {
                                        "description": "Alamat lengkap",
                                        "type": "string",
                                        "example": "Jl. Contoh No. 123"
                                    },
                                    "id_desa": {
                                        "description": "ID desa (jika ada)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil didaftarkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registrasi berhasil"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "Warga"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "description": "Endpoint untuk logout user dan menghapus JWT token.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authentication token. Dapatkan token dengan melakukan login ke endpoint /api/auth/login menggunakan username \"admintest\" dan password \"123456\"."
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}